keylogger-visualizer/requirements.txt
keyboard
matplotlib
pillow
